Migrated functions to libraries
Frequently used functions were migrated to two newly
created libraries. I settled on _pbmhelper_ as to not conflict with
with pyvmomi's _pbm_.

All updated scripts were re-tested and are working properly.

New libraries:

* `samples/tools/disk.py`
  * Functions:
    * `get_obj(content, vimtype, name)`
    * `retrieve_fcd(content, datastore, vdisk)`
    * `retrieve_fcd_snapshot(content, datastore, vdisk, snapshot)`
* `samples/tools/pbmhelper.py`
  * Functions:
    * `create_pbm_session(stub)`
    * `retrieve_storage_policy(pbmContent, policy)`
