+++ samples/attach_vdisk_to_vm.py
@@ -17,9 +17,8 @@ Python program for attaching a first class disk (fcd) to a virtual machine
 
 import atexit
 
-from tools import cli, tasks
+from tools import cli, tasks, disk
 from pyVim import connect
-from pyVim.task import WaitForTask
 from pyVmomi import vmodl
 from pyVmomi import vim
 
@@ -56,39 +55,6 @@ def get_args():
 
     my_args = parser.parse_args()
     return cli.prompt_for_password(my_args)
-
-
-def get_obj(content, vimtype, name):
-    """
-    Retrieves the managed object for the name and type specified
-    """
-    obj = None
-    container = content.viewManager.CreateContainerView(
-        content.rootFolder, vimtype, True)
-    for c in container.view:
-        if c.name == name:
-            obj = c
-            break
-    return obj
-
-
-def retrieve_fcd(content, datastore, vdisk):
-    """
-    Retrieves the managed object for the first class disk specified
-    """
-    # Set vStorageObjectManager
-    storage = content.vStorageObjectManager
-
-    # Retrieve First Class Disks
-    disk = None
-    for d in storage.ListVStorageObject(datastore):
-        disk_info = storage.RetrieveVStorageObject(d, datastore)
-        if disk_info.config.name == vdisk:
-            disk = disk_info
-            break
-    if not disk:
-        raise RuntimeError("First Class Disk not found.")
-    return disk
 
 
 def attach_fcd_to_vm(vm, vdisk, datastore):
@@ -154,10 +120,10 @@ def main():
         content = service_instance.RetrieveContent()
 
         # Retrieve Datastore Object
-        datastore = get_obj(content, [vim.Datastore], args.datastore)
+        datastore = disk.get_obj(content, [vim.Datastore], args.datastore)
 
         # Retrieve FCD Object
-        vdisk = retrieve_fcd(content, datastore, args.vdisk)
+        vdisk = disk.retrieve_fcd(content, datastore, args.vdisk)
 
         # Retrieve VM
         vm = None
@@ -165,7 +131,7 @@ def main():
             search_index = content.searchIndex
             vm = search_index.FindByUuid(None, args.uuid, True)
         elif args.vm_name:
-            vm = get_obj(content, [vim.VirtualMachine], args.vm_name)
+            vm = disk.get_obj(content, [vim.VirtualMachine], args.vm_name)
 
         # Attaching FCD to VM
         if vm:
