+++ samples/create_vdisk.py
@@ -17,9 +17,9 @@ Python program for creating a first class disk (fcd)
 
 import atexit
 
-from tools import cli, tasks
+from tools import cli, tasks, disk, pbmhelper
 from pyVim import connect
-from pyVmomi import vmodl, vim, pbm, VmomiSupport
+from pyVmomi import vmodl, vim
 
 
 def get_args():
@@ -66,80 +66,6 @@ def get_args():
     return cli.prompt_for_password(my_args)
 
 
-def get_obj(content, vimtype, name):
-    """
-    Retrieves the managed object for the name and type specified
-    """
-    obj = None
-    container = content.viewManager.CreateContainerView(
-        content.rootFolder, vimtype, True)
-    for c in container.view:
-        if c.name == name:
-            obj = c
-            break
-    return obj
-
-
-def get_pbm_connection(stub):
-    import pyVmomi
-    import ssl
-    # Make compatible with both Python2/3
-    try:
-        from http import cookies
-    except ImportError:
-        import Cookie as cookies
-
-    sessionCookie = stub.cookie.split('"')[1]
-    httpContext = VmomiSupport.GetHttpContext()
-    cookie = cookies.SimpleCookie()
-    cookie["vmware_soap_session"] = sessionCookie
-    httpContext["cookies"] = cookie
-    VmomiSupport.GetRequestContext()["vcSessionCookie"] = sessionCookie
-    hostname = stub.host.split(":")[0]
-
-    context = None
-    if hasattr(ssl, "_create_unverified_context"):
-        context = ssl._create_unverified_context()
-    pbmStub = pyVmomi.SoapStubAdapter(
-        host=hostname,
-        version="pbm.version.version1",
-        path="/pbm/sdk",
-        poolSize=0,
-        sslContext=context)
-    pbmSi = pbm.ServiceInstance("ServiceInstance", pbmStub)
-    pbmContent = pbmSi.RetrieveContent()
-
-    return pbmContent
-
-
-def retrieve_storage_policy(pbmContent, policy):
-    """
-    Retrieves the managed object for the storage policy specified
-    """
-    # Set PbmQueryProfile
-    pm = pbmContent.profileManager
-
-    # Retrieving Storage Policies
-    profileIds = pm.PbmQueryProfile(resourceType=pbm.profile.ResourceType(
-        resourceType="STORAGE"), profileCategory="REQUIREMENT"
-    )
-    if len(profileIds) > 0:
-        profiles = pm.PbmRetrieveContent(profileIds=profileIds)
-    else:
-        raise RuntimeError("No Storage Policies found.")
-
-    # Searching for Storage Policy
-    profile = None
-    for p in profiles:
-        if p.name == policy:
-            profile = p
-            break
-    if not profile:
-        raise RuntimeError("Storage Policy specified not found.")
-
-    return profile
-
-
 def main():
     """
     Simple command-line program for creating a first class disk.
@@ -164,16 +90,16 @@ def main():
         content = service_instance.RetrieveContent()
 
         # Connect to SPBM Endpoint
-        pbmContent = get_pbm_connection(service_instance._stub)
+        pbmContent = pbmhelper.create_pbm_session(service_instance._stub)
 
         # Retrieving Storage Policy
         if args.policy:
-            policy = retrieve_storage_policy(pbmContent, args.policy)
+            policy = pbmhelper.retrieve_storage_policy(pbmContent, args.policy)
         else:
             policy = None
 
         # Retrieve Datastore Object
-        datastore = get_obj(content, [vim.Datastore], args.datastore)
+        datastore = disk.get_obj(content, [vim.Datastore], args.datastore)
 
         # Setting FCD Specifications
         spec = vim.vslm.CreateSpec()
