+++ samples/list_vdisk_snapshots.py
@@ -17,7 +17,7 @@ Python program for listing all snapshots of a first class disk (fcd)
 
 import atexit
 
-from tools import cli, tasks
+from tools import cli, tasks, disk
 from pyVim import connect
 from pyVmomi import vmodl
 from pyVmomi import vim
@@ -46,42 +46,9 @@ def get_args():
     return cli.prompt_for_password(my_args)
 
 
-def get_obj(content, vimtype, name):
+def list_fcd_snapshots(content, vdisk):
     """
-    Retrieves the managed object for the name and type specified
-    """
-    obj = None
-    container = content.viewManager.CreateContainerView(
-        content.rootFolder, vimtype, True)
-    for c in container.view:
-        if c.name == name:
-            obj = c
-            break
-    return obj
-
-
-def retrieve_fcd(content, datastore, vdisk):
-    """
-    Retrieves the managed object for the first class disk specified
-    """
-    # Set vStorageObjectManager
-    storage = content.vStorageObjectManager
-
-    # Retrieve First Class Disks
-    disk = None
-    for d in storage.ListVStorageObject(datastore):
-        disk_info = storage.RetrieveVStorageObject(d, datastore)
-        if disk_info.config.name == vdisk:
-            disk = disk_info
-            break
-    if not disk:
-        raise RuntimeError("First Class Disk not found.")
-    return disk
-
-
-def retrieve_snapshots(content, vdisk):
-    """
-    Retrieves the vmware object for the snapshot specified
+    List all the snapshots for the specified first class disk
     """
     # Set vStorageObjectManager
     storage = content.vStorageObjectManager
@@ -103,7 +70,7 @@ def retrieve_snapshots(content, vdisk):
 
 def main():
     """
-    Simple command-line program for deleting a snapshot of a first class disk.
+    Simple command-line program for listing all snapshots of a fcd
     """
 
     args = get_args()
@@ -125,13 +92,13 @@ def main():
         content = service_instance.RetrieveContent()
 
         # Retrieve Datastore Object
-        datastore = get_obj(content, [vim.Datastore], args.datastore)
+        datastore = disk.get_obj(content, [vim.Datastore], args.datastore)
 
         # Retrieve FCD Object
-        vdisk = retrieve_fcd(content, datastore, args.vdisk)
+        vdisk = disk.retrieve_fcd(content, datastore, args.vdisk)
 
         # Retrieve all Snapshots
-        retrieve_snapshots(content, vdisk)
+        list_fcd_snapshots(content, vdisk)
 
     except vmodl.MethodFault as error:
         print("Caught vmodl fault : " + error.msg)
